feat: DMUserCard와 분리하면서 Friend의 UserCard도 변경
