+++ src/service/feature/friend/types/friend.ts
@@ -1,14 +1,7 @@
+import { ChannelMember } from '@service/feature/channel/types/channel';
+
 export interface FriendData {
   friendshipId: 8;
   friendshipDateTime: string;
-  friendshipInfo: FriendInfoData;
-}
-
-export interface FriendInfoData {
-  id: string;
-  nickname: string;
-  name: string;
-  avatarUrl: string;
-  state: 'ONLINE' | 'OFFLINE';
-  createdAt: string;
+  friendshipInfo: ChannelMember;
 }
